test(cf-t8k1.fu2): split lazy-init tests into 2 files per noDoMockInTestBody invariant
Melania flagged: PR #1367 test22 failed on noDoMockInTestBody (CF-fgsw).
My internationalShippingLazyInit.test.js had `vi.doMock()` calls inside
a beforeEach, which the invariant forbids (vi.doMock doesn't hoist or
invalidate static import bindings — produces silent no-op tests).

Mirroring morgott's fu1 fix pattern (commit 6355fd9 on PR #1364):

## Split

- tests/internationalShippingLazyInit.test.js — EMPTY-mock contract via
  TOP-LEVEL vi.mock. Pins:
  - imports cleanly with EMPTY public/sharedTokens mock
  - getInternationalZones throws clearly when fields missing

- tests/internationalShippingLazyInitPopulated.test.js (NEW) — POPULATED-
  mock contract via TOP-LEVEL vi.mock. Pins:
  - getInternationalZones returns the zone map
  - getRestrictedCountries returns the restricted-country list
  - getFreeInternationalThreshold returns the threshold value
  - getInternationalZones is memoized (toBe reference equality)
  - __resetCacheForTests clears all three caches

Doc comments cross-reference each other.

## Why split

Each test file must declare its mock shape ONCE at the top level. The
per-describe mock-shape swap that would otherwise require vi.doMock is
precisely what the CF-fgsw invariant forbids. Two files = two mock
shapes = no invariant violation.

## Test count unchanged (7 total)

- empty-mock file: 2 tests
- populated-mock file: 5 tests
- (was 7 in single file; same 7 split across 2 files)

## Refs

- Origin nudge: melania 2026-05-16 PR #1367 test22 FAIL
- Pattern source: morgott commit 6355fd9 on PR #1364 (fu1 equivalent)
- Invariant: tests/noDoMockInTestBody.test.js (CF-fgsw)
